Clean up 'Creating app NAME' / 'Updating app NAME' events
When pushing a single app without a manifest, we want to tell the user
whether they are creating a brand new app or they are updating an
existing app. We want this sentence to appear at the beginning or near
the beginning of the push output.

In v6, push would get an app summary to determine whether it needed to
create or update an app. In v7, without a manifest, we just attempt to
create the app and adjust the message to _Updating_ if the app already
exists.

Previous iterations of v7 push have
some "Updating app NAME" events in the middle of the flow, which we are
removing.

When pushing with a manifest, since the CLI now just gives that manifest
to the CC, the CLI no longer (compared to v6) knows whether a given app
is being created or updated, so we just say "Applying manifest..."
instead.

[#162558994]

Signed-off-by: Eric Promislow <eric.promislow@suse.com>